requirements.txt
openai>=1.0.0
streamlit>=1.55.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
ipython>=8.0.0
